Update TTL field to use long instead of double

+++ src/main/java/com/sggc/models/User.java
@@ -35,5 +35,5 @@ public class User {
     }
 
     //TTL (time-to-live) field for DynamoDB
-    private double removalDate;
+    private long removalDate;
 }
